+++ src/ic-cdk/CHANGELOG.md
@@ -6,6 +6,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.6.6] - 2022-11-09
+
+### Added
+
+- Added `StableIO` to implement both `io::Write` and `io::Read` for stable memory. (#335)
+- Added 64-bit support for `io::Write` and `io::Read` via `StableIO`.
+- Implement `io::Seek` for stable storage.
+
+### Changed
+
+-  `StableWriter` and `StableReader` are now wrappers around `StableIO`. 
+
+## [0.6.5] - 2022-11-04
+
+### Changed
+
+BREAKING CHANGE of experimental API:
+- `http_request` to support `context` field in callback function. (#326)
+
+## [0.6.4] - 2022-10-28
+
+### Added
+
+- Expose `offset` of `StableReader` and `StableWriter`. (#330)
+
+## [0.6.3] - 2022-10-26
+
+### Fixed
+
+- Doc can build on docs.rs. (#327)
+
+## [0.6.2] - 2022-10-24
+
+### Refactored
+
+- Separate `ic0` crate for system API. (#324)
+
 ## [0.6.1] - 2022-10-14
 
 ### Added